refactor: part 1 - decoupled claim settings from bond wizard
