+++ apps/extension/src/ui/domains/Portfolio/AssetDetails/PopupTokenBalances/TokenBalancesList.tsx
@@ -79,7 +79,6 @@ export const TokenBalancesList = ({
         {tokenId && (
           <div className="size-[3.8rem] shrink-0">
             <TokenContextMenu
-              balances={balances}
               tokenId={tokenId}
               className="hover:bg-grey-700 focus-visible:bg-grey-700 rounded-full"
             />
