+++ apps/extension/src/ui/domains/Staking/Bittensor/BittensorBondModal/BittensorBondFormBase.tsx
@@ -25,15 +25,15 @@ import { TokensAndFiat } from "../../../Asset/TokensAndFiat"
 import { BondAccountPicker } from "../../Bond/BondAccountPicker"
 import { SeekGetFeeDiscountsDrawer } from "../../Seek/SeekGetFeeDiscountsDrawer"
 import { STAKING_MODAL_CONTENT_CONTAINER_ID } from "../../shared/ModalContent"
+import { BittensorAssetAccountSummary } from "../components/BittensorAssetAccountSummary"
+import { BittensorStakingModalHeader } from "../components/BittensorModalHeader"
+import { BittensorModalLayout } from "../components/BittensorModalLayout"
 import { useBittensorBondModal } from "../hooks/useBittensorBondModal"
 import { ROOT_NETUID } from "../utils/constants"
 import { StakingFeeEstimate } from "./../../shared/StakingFeeEstimate"
 import { useBittensorBondWizard } from "./../hooks/useBittensorBondWizard"
 import { BittensorAvailableToUnstake } from "./BittensorAvailableToUnstake"
 import { BittensorDelegatorNameButton } from "./BittensorDelegatorNameButton"
-import { BittensorStakingModalHeader } from "./BittensorModalHeader"
-import { BittensorModalLayout } from "./BittensorModalLayout"
-import { BittensorAssetAccountSummary } from "./components/BittensorAssetAccountSummary"
 import { BittensorSelectStakeDrawer } from "./Drawers/BittensorSelectStakeDrawer"
 
 const AvailableBalance: FC<{ token: Token; account: Account }> = ({ token, account }) => {
