+++ apps/extension/src/ui/domains/Staking/Bittensor/BittensorBondModal/Forms/BittensorRootBondReview.tsx
@@ -9,9 +9,9 @@ import { SapiSendButton } from "../../../../Transactions/SapiSendButton"
 import { StakingAccountDisplay } from "../../../shared/StakingAccountDisplay"
 import { StakingFeeEstimate } from "../../../shared/StakingFeeEstimate"
 import { StakingUnbondingPeriod } from "../../../shared/StakingUnbondingPeriod"
+import { BittensorStakingModalHeader } from "../../components/BittensorModalHeader"
+import { BittensorModalLayout } from "../../components/BittensorModalLayout"
 import { useBittensorBondWizard } from "../../hooks/useBittensorBondWizard"
-import { BittensorStakingModalHeader } from "../BittensorModalHeader"
-import { BittensorModalLayout } from "../BittensorModalLayout"
 
 export const BittensorRootBondReview = () => {
   const { t } = useTranslation()
