+++ apps/extension/src/ui/domains/Staking/Bittensor/BittensorClaimSettingsModal/index.tsx
@@ -3,17 +3,17 @@ import { useCallback, useEffect, useMemo, useState } from "react"
 import { useTranslation } from "react-i18next"
 import { Button } from "talisman-ui"
 
-import type { RootClaimType } from "../../../hooks/bittensor/dTao/types"
-import { SapiSendButton } from "../../../../Transactions/SapiSendButton"
-import { BondAccountPicker } from "../../../Bond/BondAccountPicker"
-import { DEFAULT_ROOT_CLAIM_TYPE } from "../../../hooks/bittensor/dTao/types"
-import { useGetBittensorClaimType } from "../../../hooks/bittensor/dTao/useGetBittensorClaimType"
-import { useGetBittensorClaimTypePayload } from "../../../hooks/bittensor/dTao/useGetBittensorClaimTypePayload"
-import { useBittensorBondModal } from "../../hooks/useBittensorBondModal"
-import { useBittensorBondWizard } from "../../hooks/useBittensorBondWizard"
-import { BittensorStakingModalHeader } from "../BittensorModalHeader"
-import { BittensorModalLayout } from "../BittensorModalLayout"
+import type { RootClaimType } from "../../hooks/bittensor/dTao/types"
+import { SapiSendButton } from "../../../Transactions/SapiSendButton"
+import { BondAccountPicker } from "../../Bond/BondAccountPicker"
+import { DEFAULT_ROOT_CLAIM_TYPE } from "../../hooks/bittensor/dTao/types"
+import { useGetBittensorClaimType } from "../../hooks/bittensor/dTao/useGetBittensorClaimType"
+import { useGetBittensorClaimTypePayload } from "../../hooks/bittensor/dTao/useGetBittensorClaimTypePayload"
 import { BittensorAssetAccountSummary } from "../components/BittensorAssetAccountSummary"
+import { BittensorStakingModalHeader } from "../components/BittensorModalHeader"
+import { BittensorModalLayout } from "../components/BittensorModalLayout"
+import { useBittensorBondModal } from "../hooks/useBittensorBondModal"
+import { useBittensorBondWizard } from "../hooks/useBittensorBondWizard"
 
 export const BittensorClaimSettings = () => {
   const [selectedClaimType, setSelectedClaimType] = useState<RootClaimType>(DEFAULT_ROOT_CLAIM_TYPE)
