+++ apps/extension/src/ui/domains/Staking/Bittensor/components/BittensorModalHeader.tsx
