+++ apps/extension/src/ui/domains/Staking/Bittensor/components/BittensorModalLayout.tsx
